refactor: question mark icon
